requirements.txt
django
djangorestframework
drf-yasg
python-dotenv
collectstatic
django-grappelli


